Decimal

+++ src/test/java/org/libj/math/DecimalMultiplicativeOperation.java
@@ -22,7 +22,7 @@ abstract class DecimalMultiplicativeOperation extends DecimalArithmeticOperation
   }
 
   @Override
-  byte maxValuePower(final byte bits) {
+  int maxValuePower(final byte bits) {
     // For multiply and divide, the maxValue should not be so big as to cause the result to overrun max/min value
     return (byte)(bits == 6 ? 55 : bits == 5 ? 43 : bits == 4 ? 36 : bits == 3 ? 33 : bits == 2 ? 31 : 30);
   }
